pyproject.toml
[project]
name = "pytools"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "authlib>=1.3.2",
    "cadquery==2.3.0",
    "cryptography>=43.0.3",
    "fastapi[standard]>=0.115.4",
    "google-auth>=2.36.0",
    "itsdangerous>=2.2.0",
    "jupyter>=1.1.1",
    "opencv-python>=4.10.0.84",
    "pyav>=13.1.0",
    "pyjwt>=2.9.0",
    "pymupdf>=1.24.13",
    "pyparsing>=3.2.0",
    "restrictedpython>=7.4",
]

[tool.uv.workspace]
members = ["local/my_module"]
